// src/types.ts
export type ColumnIndicatorProps = {
    /** The column the indicator indicates */
    column: number;
    /** A custom label for the indicator as provided in columnLabels */
    label?: React.ReactNode | null;
    /** Whether the entire column in selected */
    selected: boolean;
    /** Currently selected point */
    selectedPoint: Point | null;
    /** Whether the cell is active */
    active: boolean;
    /** Callback to be called when the column is selected */
    onSelect: (column: number) => void;
    /** Activate the cell at the given point */
    activate: (point: Point) => void;
};

/** Type of the ColumnIndicator component */
export type ColumnIndicatorComponent = React.ComponentType<ColumnIndicatorProps>;